third_party/blink/web_tests/external/wpt/webrtc/RTCRtpTransceiver-stopping.https-expected.txt

This is a testharness.js-based test.
[FAIL] [audio] Remotely stopping a transceiver ends the track
  assert_equals: expected "live" but got "ended"
[FAIL] [audio] Rollback when removing transceiver does end the track
  assert_equals: expected "live" but got "ended"
[FAIL] [video] Remotely stopping a transceiver ends the track
  assert_equals: expected "live" but got "ended"
[FAIL] [video] Rollback when removing transceiver does end the track
  assert_equals: expected "live" but got "ended"
Harness: the test ran to completion.
